+++ cordformation/src/main/kotlin/net/corda/plugins/Dockerform.kt
@@ -1,17 +1,17 @@
 package net.corda.plugins
 
+import com.typesafe.config.ConfigFactory
+import com.typesafe.config.ConfigValueFactory
 import org.gradle.api.model.ObjectFactory
-import org.gradle.api.tasks.InputFile
-import org.gradle.api.tasks.PathSensitive
+import org.gradle.api.tasks.*
 import org.gradle.api.tasks.PathSensitivity.RELATIVE
-import org.gradle.api.tasks.TaskAction
 import org.yaml.snakeyaml.DumperOptions
 import org.yaml.snakeyaml.Yaml
-import java.nio.charset.StandardCharsets
 import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.Paths
 import javax.inject.Inject
+
 
 /**
  * Creates docker-compose file and image definitions based on the configuration of this task in the gradle configuration DSL.
@@ -22,9 +22,10 @@ import javax.inject.Inject
 open class Dockerform @Inject constructor(objects: ObjectFactory) : Baseform(objects) {
 
     private companion object {
+
+        private const val DEFAULT_DB_STARTING_PORT = 5432
         private const val DEFAULT_SSH_PORT = 22022
         private val DEFAULT_DIRECTORY: Path = Paths.get("build", "docker")
-
         private const val COMPOSE_SPEC_VERSION = "3"
 
         private val YAML_FORMAT_OPTIONS = DumperOptions().apply {
@@ -41,7 +42,6 @@ open class Dockerform @Inject constructor(objects: ObjectFactory) : Baseform(obj
 
     private val directoryPath: Path = project.projectDir.toPath().resolve(directory)
 
-
     val dockerComposePath: Path
         @PathSensitive(RELATIVE)
         @InputFile
@@ -54,6 +54,17 @@ open class Dockerform @Inject constructor(objects: ObjectFactory) : Baseform(obj
             return wantedPath
         }
 
+    @get:Optional
+    @get:Input
+    var dockerConfig: Map<String, Any> = emptyMap()
+
+    /**
+     * Docker image for this node, or null if one hasn't been specified.
+     */
+    @get:Optional
+    @get:Input
+    var dockerImage: String? = null
+
     /**
      * This task action will create and install the nodes based on the node configurations added.
      */
@@ -61,20 +72,20 @@ open class Dockerform @Inject constructor(objects: ObjectFactory) : Baseform(obj
     fun build() {
         project.logger.lifecycle("Running DockerForm task")
         initializeConfiguration()
-        nodes.forEach { it -> it.installDockerConfig(DEFAULT_SSH_PORT) }
+        nodes.forEach { it.installDockerConfig(DEFAULT_SSH_PORT) }
         installCordaJar()
-        nodes.forEach(Node::installDrivers)
         generateKeystoreAndSignCordappJar()
         generateExcludedWhitelist()
         bootstrapNetwork()
         nodes.forEach(Node::buildDocker)
 
+        val services = mutableMapOf<String, Map<String, Any>>()
 
-        // Transform nodes path the absolute ones
-        val services = nodes.map {
+        nodes.forEachIndexed {index, it ->
+
             val nodeBuildDir = directoryPath.resolve(it.nodeDir.name).toAbsolutePath().toString()
 
-            it.containerName to mapOf(
+            val service = mutableMapOf(
                     "volumes" to listOf(
                             "$nodeBuildDir/node.conf:/etc/corda/node.conf",
                             "$nodeBuildDir/certificates:/opt/corda/certificates",
@@ -84,13 +95,79 @@ open class Dockerform @Inject constructor(objects: ObjectFactory) : Baseform(obj
                             "$nodeBuildDir/network-parameters:/opt/corda/network-parameters",
                             "$nodeBuildDir/additional-node-infos:/opt/corda/additional-node-infos",
                             "$nodeBuildDir/drivers:/opt/corda/drivers"
-
                     ),
                     "ports" to listOf(it.rpcPort, it.config.getInt("sshd.port")),
-                    "image" to "corda/corda-zulu-${it.runtimeVersion().toLowerCase()}"
+                    "image" to (dockerImage ?: "corda/corda-zulu-${it.runtimeVersion().toLowerCase()}")
             )
-        }.toMap()
 
+            if (dockerConfig.isNotEmpty()) {
+
+                val dockerConfig = ConfigFactory.parseMap(dockerConfig)
+
+                val dbDockerfile = dockerConfig.getString("dockerConfig.dbDockerfile")
+                val dbDockerfileArgs = dockerConfig.getConfig("dockerConfig.dbDockerfileArgs")
+                val dbUser = dockerConfig.getString("dockerConfig.dbUser")
+                val dbPassword = dockerConfig.getString("dockerConfig.dbPassword")
+                val dbDataSourceClassName = dockerConfig.getString("dataSourceProperties.dataSourceClassName")
+                val dbTransactionIsolationLevel = dockerConfig.getString("database.transactionIsolationLevel")
+                val dbRunMigration = dockerConfig.getBoolean("database.runMigration")
+                val dbSchema= dockerConfig.getString("dockerConfig.dbSchema")
+
+                // These are allocated by docker-compose
+                val dbPort = DEFAULT_DB_STARTING_PORT + index
+                val dbHost = "${it.containerName}-db"
+                val defaultUrlArgs = ConfigFactory.empty()
+                        .withValue("DBHOSTNAME",ConfigValueFactory.fromAnyRef(dbHost))
+                        .withValue("DBPORT",ConfigValueFactory.fromAnyRef(dbPort))
+
+                // Override the port and hostname
+                val dockerfileArgs = defaultUrlArgs.withFallback(dbDockerfileArgs).resolve()
+
+                var dbUrl = dockerConfig.getString("dataSourceProperties.dataSource.url")
+
+                if (dockerConfig.hasPath("dataSourceProperties.dataSource.urlArgs")) {
+                    val urlArgs = defaultUrlArgs.withFallback(
+                            dockerConfig.getConfig("dataSourceProperties.dataSource.urlArgs")).resolve()
+
+                    val dbUrlConfig = ConfigFactory.parseString("DBURL=${TypesafeUtils.encodeString(dbUrl)}")
+                            .resolveWith(urlArgs)
+
+                    dbUrl = dbUrlConfig.getString("DBURL")
+                }
+
+                val dbConfig = ConfigFactory.empty()
+                        .withValue("dataSourceProperties.dataSourceClassName", ConfigValueFactory.fromAnyRef(dbDataSourceClassName))
+                        .withValue("dataSourceProperties.dataSource.url", ConfigValueFactory.fromAnyRef(dbUrl))
+                        .withValue("dataSourceProperties.dataSource.user", ConfigValueFactory.fromAnyRef(dbUser))
+                        .withValue("dataSourceProperties.dataSource.password", ConfigValueFactory.fromAnyRef(dbPassword))
+                        .withValue("database.transactionIsolationLevel", ConfigValueFactory.fromAnyRef(dbTransactionIsolationLevel))
+                        .withValue("database.runMigration", ConfigValueFactory.fromAnyRef(dbRunMigration))
+                        .withValue("database.schema", ConfigValueFactory.fromAnyRef(dbSchema))
+
+                it.installDefaultDatabaseConfig(dbConfig)
+
+                val args = mutableMapOf<String, Any>()
+
+                dockerfileArgs.entrySet().toList().map {
+                    args[it.key] = it.value.unwrapped()
+                }
+
+                services[dbHost] = mapOf(
+                        "build" to mapOf(
+                             "context" to project.buildDir.absolutePath,
+                             "dockerfile" to project.buildDir.resolve(dbDockerfile).toString(),
+                             "args" to args
+                        ),
+                        "restart" to "unless-stopped",
+                        "ports" to listOf(dbPort)
+                )
+
+                service["image"] = dockerImage ?: "entdocker.software.r3.com/corda-enterprise-java1.8-${it.runtimeVersion().toLowerCase()}"
+                service["depends_on"] = listOf(dbHost)
+            }
+
+            services[it.containerName] = service
+        }
 
         val dockerComposeObject = mapOf(
                 "version" to COMPOSE_SPEC_VERSION,
@@ -98,7 +175,6 @@ open class Dockerform @Inject constructor(objects: ObjectFactory) : Baseform(obj
 
         val dockerComposeContent = YAML_MAPPER.dump(dockerComposeObject)
 
-        Files.write(dockerComposePath, dockerComposeContent.toByteArray(StandardCharsets.UTF_8))
+        Files.write(dockerComposePath, dockerComposeContent.toByteArray())
     }
 }
-