+++ cordformation/src/test/kotlin/net/corda/plugins/BaseformTest.kt
@@ -36,7 +36,9 @@ open class BaseformTest {
     fun getStandardGradleRunnerFor(buildFileResourceName: String, taskName: String = "deployNodes"): GradleRunner {
         createBuildFile(buildFileResourceName)
         installResource("settings.gradle")
+        installResource("repositories.gradle")
         installResource("gradle.properties")
+        installResource("postgres.gradle")
         return GradleRunner.create()
                 .withDebug(true)
                 .withProjectDir(testProjectDir.toFile())